refactor(cortes): split 1260-LOC page into focused module
Extracts the RDB cortes de caja dashboard from a single-file
app/rdb/cortes/page.tsx into components/cortes/, following the same
convention used by components/app-shell/, components/health/,
components/travel/, components/rh/, and components/tasks/.

Behavior preserved 1:1 — no SQL, data-fetching, calculation, or UI
change. Production-critical module (Beto checks cortes daily), so
every split keeps structure/props identical to what the orchestrator
previously rendered inline.

Each file < 325 LOC; page.tsx becomes a thin wrapper around
<CortesView />. Typecheck, lint, prettier, and vitest (131/131) pass.
